requirements.txt
Flask>=3.0.0
openai-whisper>=20231117
Flask-Login>=0.6.3
python-dotenv>=1.0.1
firebase-admin>=6.5.0
gunicorn>=21.2.0
